Fixes errors, cleanup, and the first successful frame.
