+++ cpp-foundation/src/main.cpp
@@ -1,287 +1,23 @@
-#include <iostream>
-#include <chrono>
-#include <iomanip>
-#include <vector>
 #include <fstream>
-#include <memory>
-
-// Include your electromagnetic spectrum foundations
-// (These will include from cpp-foundation/include/ automatically)
+#include <algorithm>
 #include "static_frame_generator.hpp"
-#include "intelligent_term_delegator.hpp"
-#include "precision_safe_threading.hpp"
-#include "context_overflow_guard.hpp"
-#include "adaptive_gpu_delegation.hpp"
-
-// Optional includes if they exist
-#ifdef CORTEX_BASELINE_AVAILABLE
-#include "cortex_baseline_processor.hpp"
-#include "cortex_buffer_baseline.hpp"
-#endif
-
-using namespace cortex::frames;
-
-using CosmicPrecision = boost::multiprecision::cpp_dec_float<CORTEX_EM_SPECTRUM_PRECISION>;
-
-void print_project_header() {
-    std::cout << "🌈 CORTEX ELECTROMAGNETIC SPECTRUM FOUNDATION\n";
-    std::cout << "==================================================================\n";
-    std::cout << "Project: cortex-em-spectrum-foundation\n";
-    std::cout << "Python Optimization Foundations Successfully Translated to C++:\n";
-    std::cout << "💎 Dynamic Optimized Library → Treasure Vault (Memory Optimization)\n";
-    std::cout << "🤯 Context Overflow Guard → Recursive Self-Capturing Protection\n";
-    std::cout << "🧠 Intelligent Term Delegator → Complexity-Based Routing\n";
-    std::cout << "🎯 Precision Safe Threading → Zero-Drift 141-Decimal Threading\n";
-    std::cout << "🚀 Adaptive GPU Delegation → Cross-Architecture Optimization\n";
-    std::cout << "🖼️ Static Frame Generator → Electromagnetic Spectrum Processing\n";
-    std::cout << "==================================================================\n\n";
-}
-
-void test_project_organization() {
-    std::cout << "🔧 PROJECT ORGANIZATION VALIDATION\n";
-    std::cout << "-" << std::string(60, '-') << "\n";
-
-    std::cout << "   ✅ Headers: cpp-foundation/include/\n";
-    std::cout << "   ✅ Sources: cpp-foundation/src/\n";
-    std::cout << "   ✅ Python Documentation: optimization-scripts/\n";
-    std::cout << "   ✅ Visual Studio Project: cortex_foundation.vcxproj\n";
-    std::cout << "   ✅ CMake Support: CMakeLists.txt\n";
-    std::cout << "   ✅ Project Structure: IEEE SOFTWARE ENGINEERING STANDARD\n";
-    std::cout << "   ✅ No Python dependencies for C++ compilation\n";
-    std::cout << "   ✅ Python treasures preserved as documentation\n\n";
-}
-
-void test_electromagnetic_spectrum_processing() {
-    std::cout << "🌈 ELECTROMAGNETIC SPECTRUM PROCESSING TEST\n";
-    std::cout << "-" << std::string(60, '-') << "\n";
-
-    try {
-        // Initialize your electromagnetic spectrum foundation
-        StaticFrameGenerator generator;
-
-        std::cout << "🖼️ Static Frame Generator initialized successfully\n";
-
-        // Test wavelength to RGB conversion
-        std::vector<std::pair<CosmicPrecision, std::string>> test_wavelengths = {
-            {CosmicPrecision("380"), "Violet"},
-            {CosmicPrecision("450"), "Blue"},
-            {CosmicPrecision("500"), "Green"},
-            {CosmicPrecision("580"), "Yellow"},
-            {CosmicPrecision("650"), "Red"},
-            {CosmicPrecision("750"), "Near-infrared"}
-        };
-
-        std::cout << "🌈 Testing wavelength to RGB conversion:\n";
-        for (const auto& [wavelength, color] : test_wavelengths) {
-            auto pixel = generator.wavelength_to_rgb_pixel(wavelength);
-            auto intensity = generator.wavelength_to_rgb_intensity(wavelength);
-
-            std::cout << "   " << wavelength << "nm (" << std::setw(12) << color << "): ";
-            std::cout << "R=" << std::fixed << std::setprecision(3) << pixel.red;
-            std::cout << " G=" << pixel.green;
-            std::cout << " B=" << pixel.blue;
-            std::cout << " I=" << intensity << "\n";
-        }
-
-        std::cout << "\n✅ Electromagnetic spectrum conversion test completed\n\n";
-
-    } catch (const std::exception& e) {
-        std::cout << "❌ Electromagnetic spectrum test error: " << e.what() << "\n\n";
-    }
-}
-
-void test_frame_generation() {
-    std::cout << "🖼️ STATIC FRAME GENERATION TEST\n";
-    std::cout << "-" << std::string(60, '-') << "\n";
-
-    try {
-        StaticFrameGenerator generator;
-
-        auto start_time = std::chrono::high_resolution_clock::now();
-
-        // Set frame intensity parameters
-        generator.set_intensity_scale(CosmicPrecision("0.6")); // moderate brightness
-
-        // Generate test electromagnetic spectrum frame
-        auto test_frame = generator.generate_test_frame(40, 30);
-
-        auto end_time = std::chrono::high_resolution_clock::now();
-        auto duration = std::chrono::duration<double>(end_time - start_time).count();
-
-        std::cout << "📊 Frame Generation Results:\n";
-        std::cout << "   Frame dimensions: " << test_frame.width << "x" << test_frame.height << "\n";
-        std::cout << "   Pixels generated: " << test_frame.pixels.size() << "\n";
-        std::cout << "   Spectrum range: " << test_frame.spectrum_range << "\n";
-        std::cout << "   Total energy: " << test_frame.total_energy << "\n";
-        std::cout << "   Generation time: " << std::fixed << std::setprecision(6) << duration << "s\n";
-        std::cout << "   Throughput: " << std::fixed << std::setprecision(0)
-                  << (test_frame.pixels.size() / duration) << " pixels/second\n";
-
-        // Save frame data
-        generator.save_frame_data(test_frame, "electromagnetic_spectrum_test_400x300.dat");
-
-        std::cout << "✅ Static frame generation test completed\n\n";
-
-    } catch (const std::exception& e) {
-        std::cout << "❌ Frame generation test error: " << e.what() << "\n\n";
-    }
-}
-
-void test_optimization_foundations() {
-    std::cout << "⚡ OPTIMIZATION FOUNDATIONS INTEGRATION TEST\n";
-    std::cout << "-" << std::string(60, '-') << "\n";
-
-    try {
-        std::cout << "🔍 Testing optimization components integration:\n";
-
-        // Test data creation
-        std::vector<CosmicPrecision> test_data;
-        for (int i = 0; i < 1000; ++i) {
-            test_data.push_back(CosmicPrecision(std::to_string(380 + (i * 0.37))));
-        }
-
-        std::cout << "   ✅ Test data created: " << test_data.size() << " wavelength values\n";
-        std::cout << "   ✅ Range: 380nm - 750nm (visible spectrum)\n";
-        std::cout << "   ✅ Intelligent Term Delegator: Ready for complexity analysis\n";
-        std::cout << "   ✅ Precision Safe Threading: Ready for zero-drift processing\n";
-        std::cout << "   ✅ Context Overflow Guard: Ready for recursive protection\n";
-        std::cout << "   ✅ Adaptive GPU Delegation: Ready for hardware optimization\n";
-        std::cout << "   ✅ Static Frame Generator: Ready for visual conversion\n";
-
-        // Performance baseline test
-        auto start_time = std::chrono::high_resolution_clock::now();
-
-        StaticFrameGenerator generator;
-        int processed_count = 0;
-
-        for (int i = 0; i < 100; ++i) {
-            auto wavelength = test_data[i];
-            auto pixel = generator.wavelength_to_rgb_pixel(wavelength);
-            processed_count++;
-        }
-
-        auto end_time = std::chrono::high_resolution_clock::now();
-        auto duration = std::chrono::duration<double>(end_time - start_time).count();
-
-        std::cout << "\n📊 Optimization Performance Baseline:\n";
-        std::cout << "   Processed wavelengths: " << processed_count << "\n";
-        std::cout << "   Processing time: " << std::fixed << std::setprecision(6) << duration << "s\n";
-        std::cout << "   Throughput: " << std::fixed << std::setprecision(0)
-                  << (processed_count / duration) << " wavelengths/second\n";
-
-        std::cout << "\n✅ All optimization foundations integrated successfully\n\n";
-
-    } catch (const std::exception& e) {
-        std::cout << "❌ Optimization foundations test error: " << e.what() << "\n\n";
-    }
-}
-
-void generate_sample_output() {
-    std::cout << "🎨 SAMPLE OUTPUT GENERATION\n";
-    std::cout << "-" << std::string(60, '-') << "\n";
-
-    try {
-        StaticFrameGenerator generator;
-
-        // Generate sample electromagnetic spectrum frame
-        auto sample_frame = generator.generate_test_frame(200, 150);
-
-        // Generate simple PPM file for viewing
-        std::string filename = "electromagnetic_spectrum_sample.ppm";
-        std::ofstream ppm_file(filename);
-
-        if (ppm_file.is_open()) {
-            // PPM header
-            ppm_file << "P3\n";
-            ppm_file << sample_frame.width << " " << sample_frame.height << "\n";
-            ppm_file << "255\n";
-
-            // Convert pixels to RGB values
-            for (const auto& pixel : sample_frame.pixels) {
-                int r = static_cast<int>(std::min(255.0, 255.0 * static_cast<double>(pixel.red)));
-                int g = static_cast<int>(std::min(255.0, 255.0 * static_cast<double>(pixel.green)));
-                int b = static_cast<int>(std::min(255.0, 255.0 * static_cast<double>(pixel.blue)));
-
-                ppm_file << r << " " << g << " " << b << " ";
-            }
-
-            ppm_file.close();
-
-            std::cout << "🎨 Sample PPM file generated: " << filename << "\n";
-            std::cout << "   You can view this file with image viewers that support PPM format\n";
-            std::cout << "   This shows the electromagnetic spectrum as a visual gradient!\n";
-        } else {
-            std::cout << "❌ Could not create PPM file\n";
-        }
-
-        std::cout << "✅ Sample output generation completed\n\n";
-
-    } catch (const std::exception& e) {
-        std::cout << "❌ Sample output generation error: " << e.what() << "\n\n";
-    }
-}
-
-void run_complete_foundation_tests() {
-    print_project_header();
-
-    std::cout << "🚀 Running complete electromagnetic spectrum foundation tests...\n\n";
-
-    auto total_start_time = std::chrono::high_resolution_clock::now();
-
-    // Run all tests
-    test_project_organization();
-    test_electromagnetic_spectrum_processing();
-    test_frame_generation();
-    test_optimization_foundations();
-    generate_sample_output();
-
-    auto total_end_time = std::chrono::high_resolution_clock::now();
-    auto total_duration = std::chrono::duration<double>(total_end_time - total_start_time).count();
-
-    std::cout << "📊 COMPLETE FOUNDATION TEST SUMMARY\n";
-    std::cout << "========================================\n";
-    std::cout << "Total testing time: " << std::fixed << std::setprecision(3)
-              << total_duration << " seconds\n";
-    std::cout << "Project structure: IEEE SOFTWARE ENGINEERING STANDARD ✅\n";
-    std::cout << "Python foundations: Preserved as documentation ✅\n";
-    std::cout << "C++ implementation: Professional & complete ✅\n";
-    std::cout << "Electromagnetic spectrum: Processing ready ✅\n";
-    std::cout << "Visual output: Generated successfully ✅\n\n";
-}
 
 int main() {
-    try {
-        run_complete_foundation_tests();
+    cortex::StaticFrameGenerator gen;
 
-        std::cout << "🎉 CORTEX ELECTROMAGNETIC SPECTRUM FOUNDATION COMPLETE!\n\n";
-        std::cout << "🏴‍☠️ YOUR PYTHON OPTIMIZATION TREASURES HAVE BEEN SUCCESSFULLY\n";
-        std::cout << "   TRANSLATED TO PROFESSIONAL C++ ELECTROMAGNETIC SPECTRUM\n";
-        std::cout << "   PROCESSING WITH 141-DECIMAL PRECISION!\n\n";
+    const size_t width = 40, height = 30;
+    const size_t spp_x = 4, spp_y = 4; // 16 samples per pixel
+    auto frame = gen.generate_supersampled_frame(width, height, spp_x, spp_y, true);
 
-        std::cout << "📁 Project Summary:\n";
-        std::cout << "   Structure: cortex-em-spectrum-foundation/\n";
-        std::cout << "   C++ Headers: cpp-foundation/include/\n";
-        std::cout << "   C++ Sources: cpp-foundation/src/\n";
-        std::cout << "   Python Documentation: optimization-scripts/\n";
-        std::cout << "   Visual Studio Project: cortex_foundation.vcxproj\n";
-        std::cout << "   Output Files: Generated in working directory\n\n";
+    std::ofstream ppm("electromagnetic_spectrum_supersampled.ppm");
+    if (!ppm.is_open()) return 1;
 
-        std::cout << "🌈 READY FOR:\n";
-        std::cout << "   ⚡ CudaC++ Phase Calculations\n";
-        std::cout << "   🔧 Performance Optimization\n";
-        std::cout << "   📈 5000-Decimal Precision Scaling\n";
-        std::cout << "   🖼️ Advanced Visual Processing\n";
-        std::cout << "   💰 Commercial Optimization Services\n\n";
-
-        std::cout << "🍲 Soup-powered electromagnetic spectrum mastery achieved!\n";
-
-        return 0;
-
-    } catch (const std::exception& e) {
-        std::cerr << "❌ Critical error: " << e.what() << std::endl;
-        return 1;
-    } catch (...) {
-        std::cerr << "❌ Unknown critical error occurred" << std::endl;
-        return 1;
+    ppm << "P3\n" << frame.width << " " << frame.height << "\n255\n";
+    for (const auto& p : frame.pixels) {
+        int r = static_cast<int>(std::min(255.0, 255.0 * static_cast<double>(p.red)));
+        int g = static_cast<int>(std::min(255.0, 255.0 * static_cast<double>(p.green)));
+        int b = static_cast<int>(std::min(255.0, 255.0 * static_cast<double>(p.blue)));
+        ppm << r << ' ' << g << ' ' << b << '\n';
     }
+    return 0;
 }
